class2/pyproject.toml
[project]
name = "hello-world-ai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.4",
    "pillow>=11.1.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.1",
    "speechrecognition>=3.14.1",
    "streamlit>=1.43.2",
]
